Added parameterization to use libtool.
If LIBTOOL is defined to have a value, then use it to compile
and link executables.

Re-write ".o" to ".lo" in all output && dependencies, so
that libtool works.

Re-write any output file from ".a" or ".so" to ".la".

Re-write any pre-requisites from ".a" or ".so" to ".la".

This lets the user write their own rules to build the programs,
and then simply define LIBTOOL to get cross-platform build support.