Allow TARGET to be empty, which means "don't build it"
This allows a "configure" stage to set TARGET to be empty if
the pre-requisites are not avilable on a particular platform.

If TARGET is *not* defined in a submakefile, then the previous
target is used, as before.  If the previous target wasn't being
built, the the SOURCES in the submakefile are ignored, too.