Simplify the process of finding the top-level directory.
If we're run from the top-level directory, use that.  If we're run
via "include ../../../../Makefile" from a subdirectory, figure that out
without walking up the directory stack.  Instead, we look at the name
of the current Makefile.  If it's "../../../Makefile", we go back up
to the root directory, and run "make" there.