Changes to work with libtool.
When LIBTOOL is defined, ".a" and ".so" are automatically re-written
to ".la".  Libtool is used for compile / link / install, and most
commonly used options are added so it "just works".

However, you MUST currently define LIBTOOL prior to including "main.mk".
this is because the ADD_TARGET_RULE* defs get re-written when LIBTOOL
is define.  For now, this was easier to do than making only one version
of those defs that "does the right thing".

It's probably not hard, but not something I have time for right now.