Add ".libs/" in the correct place.
The previous code added ".libs/" to the *start* of a pathname.  If the
path contained a "/", this resulted in "foo/bar.lo -> .libs/foo/bar.lo",
which was wrong.  It should instead map "foo/bar.lo -> foo/.libs/bar.lo"

This is especially important when the path contains "../", as with
the boilermake subdirectory support.

There may be cases (e.g. cygwin) where the .libs directory isn't added
correctly.  But they should now be easy to fix.