Add top-level targets for everything.
This lets us be in ANY subdirectory, and do "make libplants.a".
It will then do the Right Thing to build the target.

We also add non-libtool targets, so that we can do "make libplants.la"
AND "make libplants.a" when libtool is defined.  This lets external
build systems behave (mostly) the same when libtool is defined.