Split TGT_PREREQS into pre-requisites and libraries.
This simplifies the application Makefiles.  Instead of adding explicit
LDFLAGS and LDLIBS to link to dependencies, they just declare the
dependencies.  Boilermake then takes care of filtering the dependencies
to discover which ones are libraries, and automatically adds link
lines for those libraries.

For now, the link link just links the library in as if it's an object
file.  If this is a problem later, we can add a TGT_PRLDFLAGS macro,
which filters the library dependencies to set -L and -l properly.
The ADD_TARGET_RULE.exe macro can then be updated to depend on PRLIBS,
but us PRLDFLAGS in the link line.