Use LIBTOOL=JLIBTOOL, instead of full path
We'd like to be friendly to the users, and have them *not* remember
paths to libtool.  So... specifying the magic JLIBTOOl flag means
that boilermake will figure it out.

Also, jlibtool now goes into the ${BUILD_DIR}/make/ directory.
If you have multiple versions of libtool, each one is now target-dependent