First stab at adding relink rules.
When building the targets, we use "rpath" so that the local executables
will work correctly.  However, this means that the executables depend
on the libraries in the build directory.

When we install the executables, we need to re-link them using an
"rpath" that points to the installed libraries.  This means adding
"relink" rules for all libraries and executables.